requirements.txt
requests
psutil
PYjwt
protobuf
httpx
urllib3
pycryptodome